Analytics: Split encounter identification and encounter data
